requirements.txt
azure-identity==1.18.0
azure-ai-ml==1.20.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==26.0.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.5.0
azure-storage-file-share==12.18.0
azureml-core==1.57.0.post1
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.4
azureml-fsspec==1.3.1
azureml-mlflow==1.57.0.post1

pydicom>2.4.0,<3.0.0
pandas<2.0.0
torchxrayvision~=1.3.0


pytorch-lightning<2.0.0